+++ src/local-machine.hpp
@@ -21,7 +21,6 @@
 #include <optional>
 #include <string>
 
-#include "access-log.hpp"
 #include "address-range-description.hpp"
 #include "back-merkle-tree.hpp"
 #include "hash-tree-proof.hpp"
@@ -33,6 +32,7 @@
 #include "machine-hash.hpp"
 #include "machine-runtime-config.hpp"
 #include "mcycle-root-hashes.hpp"
+#include "step-log-data.hpp"
 #include "uarch-cycle-root-hashes.hpp"
 #include "uarch-interpret.hpp"
 
@@ -59,13 +59,13 @@ private:
     mcycle_root_hashes do_collect_mcycle_root_hashes(uint64_t mcycle_end, uint64_t log2_mcycle_period,
         uint64_t mcycle_phase, int32_t log2_bundle_mcycle_count,
         const std::optional<back_merkle_tree> &previous_partial_bundle) override;
-    interpreter_break_reason do_log_step(uint64_t mcycle_count, const std::string &filename) override;
+    log_step_result do_log_step(uint64_t mcycle_count) override;
     void do_store(const std::string &directory, sharing_mode sharing) const override;
     void do_clone_stored(const std::string &from_dir, const std::string &to_dir) const override;
     void do_rename_stored(const std::string &from_dir, const std::string &to_dir) const override;
     void do_remove_stored(const std::string &dir) const override;
     void do_sync_stored(const std::string &dir) const override;
-    access_log do_log_step_uarch(const access_log::type &log_type) override;
+    log_step_uarch_result do_log_step_uarch(uint64_t uarch_cycle_count) override;
     hash_tree_proof do_get_proof(uint64_t address, int log2_target_size, int log2_root_size) const override;
     machine_hash do_get_root_hash() const override;
     machine_hash do_read_revert_root_hash() const override;
@@ -90,7 +90,7 @@ private:
     void do_set_runtime_config(const machine_runtime_config &r) override;
     void do_destroy() override;
     void do_reset_uarch() override;
-    access_log do_log_reset_uarch(const access_log::type &log_type) override;
+    step_log_data do_log_reset_uarch() override;
     uarch_interpreter_break_reason do_run_uarch(uint64_t uarch_cycle_end) override;
     uarch_cycle_root_hashes do_collect_uarch_cycle_root_hashes(uint64_t mcycle_end,
         int32_t log2_bundle_uarch_cycle_count, const machine_hashes &revert_uarch_tail) override;
@@ -98,17 +98,19 @@ private:
     machine_cmio_request do_receive_cmio_request(std::span<uint8_t> data) const override;
     void do_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
         std::optional<const_machine_hash_view> revert_root_hash) override;
-    access_log do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
-        const_machine_hash_view revert_root_hash, const access_log::type &log_type) override;
+    step_log_data do_log_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
+        const_machine_hash_view revert_root_hash) override;
     uint64_t do_get_reg_address(reg r) const override;
     machine_config do_get_default_config() const override;
     std::string do_get_address_name(uint64_t paddr) const override;
-    machine_hash do_verify_step(const_machine_hash_view root_hash_before, const std::string &log_filename,
+    machine_hash do_verify_step(const_machine_hash_view root_hash_before, std::span<const unsigned char> log,
         uint64_t mcycle_count) const override;
-    machine_hash do_verify_step_uarch(const_machine_hash_view root_hash_before, const access_log &log) const override;
-    machine_hash do_verify_reset_uarch(const_machine_hash_view root_hash_before, const access_log &log) const override;
+    machine_hash do_verify_step_uarch(const_machine_hash_view root_hash_before, std::span<const unsigned char> log,
+        uint64_t uarch_cycle_count) const override;
+    machine_hash do_verify_reset_uarch(const_machine_hash_view root_hash_before,
+        std::span<const unsigned char> log) const override;
     machine_hash do_verify_send_cmio_response(uint16_t reason, const unsigned char *data, uint64_t length,
-        const_machine_hash_view root_hash_before, const access_log &log,
+        const_machine_hash_view root_hash_before, std::span<const unsigned char> log,
         const_machine_hash_view revert_root_hash) const override;
 
     machine *get_machine();