+++ src/uarch-reset-state.hpp
@@ -26,20 +26,17 @@ template <typename STATE_ACCESS>
 void uarch_reset_state(STATE_ACCESS &a);
 
 class uarch_state_access;
-class uarch_record_state_access;
-class uarch_replay_state_access;
+class uarch_record_step_state_access;
+struct no_step_log_dumper;
+class step_log_dumper;
+template <typename Dumper>
+class uarch_replay_step_state_access;
 
-// Declaration of explicit instantiation in module uarch-reset-state.cpp
+// Declaration of explicit instantiations in module uarch-reset-state.cpp
 extern template void uarch_reset_state(uarch_state_access &a);
-
-// Declaration of explicit instantiation in module uarch-reset-state.cpp
-extern template void uarch_reset_state(uarch_state_access &a);
-
-// Declaration of explicit instantiation in module uarch-reset-state.cpp
-extern template void uarch_reset_state(uarch_record_state_access &a);
-
-// Declaration of explicit instantiation in module uarch-reset-state.cpp
-extern template void uarch_reset_state(uarch_replay_state_access &a);
+extern template void uarch_reset_state(uarch_record_step_state_access &a);
+extern template void uarch_reset_state(uarch_replay_step_state_access<no_step_log_dumper> &a);
+extern template void uarch_reset_state(uarch_replay_step_state_access<step_log_dumper> &a);
 
 } // namespace cartesi