+++ tests/lua/run-rv64i-arch-test.lua
@@ -27,7 +27,7 @@ Usage:
 
 where:
   <uarch-ram-image>
-  name of file containing the image of microemulator RAM.
+  name of file containing the image of uarch RAM.
 
   <output-signature-file>
   name of file to write test signature results
@@ -54,10 +54,10 @@ local config = {
 
 local machine = assert(cartesi.machine(config))
 
--- run microarchitecture
+-- run uarch
 machine:run_uarch()
 
--- extract test result signature from microarchitecture RAM
+-- extract test result signature from uarch RAM
 local mem = machine:read_memory(uarch_ram_start, cartesi.UARCH_RAM_LENGTH)
 local _, e1 = string.find(mem, "BEGIN_CTSI_SIGNATURE____")
 local s2, _ = string.find(mem, "END_CTSI_SIGNATURE______")